Add SelectWorkspace command
A user can now select the workspace for the discord server. The command will trigger a response to the message with a dropdown of all workspaces the user as a part of. When the user has selected one from the list, the current workspace is set for that server.

This feature includes a Workspace model to hold all attrtibutes of a model and also implement some static methods to communicate with the api.